Update payment model and request types to use float64 for Amount field

+++ internal/models/payment.go
@@ -20,7 +20,7 @@ type Payment struct {
 	BaseModel
 	OrderID        string        `gorm:"index"`
 	IdempotencyKey string        `gorm:"uniqueIndex:uniq_idem_key"`
-	Amount         int64         `gorm:"not null"`
+	Amount         float64       `gorm:"not null"`
 	Status         PaymentStatus `gorm:"type:text;index;not null"`
 	Attempts       int           `gorm:"default:0"`
 	LastError      string        `gorm:"type:text"`
@@ -31,16 +31,16 @@ func (Payment) TableName() string {
 }
 
 type CreatePaymentRequest struct {
-	OrderID        string `json:"order_id" binding:"required"`
-	IdempotencyKey string `json:"idempotency_key" binding:"required"`
-	Amount         int64  `json:"amount" binding:"required"`
+	OrderID        string  `json:"order_id" binding:"required"`
+	IdempotencyKey string  `json:"idempotency_key" binding:"required"`
+	Amount         float64 `json:"amount" binding:"required"`
 }
 
 type CreatePaymentResponseData struct {
 	PaymentID      uuid.UUID `json:"payment_id"`
 	OrderID        string    `json:"order_id"`
 	IdempotencyKey string    `json:"idempotency_key"`
-	Amount         int64     `json:"amount"`
+	Amount         float64   `json:"amount"`
 	Status         string    `json:"status"`
 }
